SKFP-223: Rename file
